# docs/swagger.yaml
basePath: /
definitions:
  rmbproxy.Message:
    properties:
      cmd:
        example: zos.statistics.get
        type: string
      dat:
        example: ""
        type: string
      dst:
        example:
        - 2
        - 3
        items:
          type: integer
        type: array
      err:
        example: ""
        type: string
      exp:
        example: 0
        type: integer
      now:
        example: 1631078674
        type: integer
      ret:
        example: ""
        type: string
      shm:
        example: ""
        type: string
      src:
        example: 1
        type: integer
      try:
        example: 2
        type: integer
      uid:
        example: ""
        type: string
      ver:
        example: 1
        type: integer
    type: object
  rmbproxy.MessageIdentifier:
    properties:
      retqueue:
        example: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx
        type: string
    type: object
...
info:
  contact: {}
  description: grid proxy server has the main methods to list farms, nodes, node details
    in the grid.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Grid Proxy Server API
  version: "1.0"
paths:
...
  /twin/{twin_id}:
    post:
      consumes:
      - application/json
      description: submit the message
      parameters:
      - description: rmb.Message
        in: body
        name: msg
        required: true
        schema:
          $ref: '#/definitions/rmbproxy.Message'
      - description: twin id
        in: path
        name: twin_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rmbproxy.MessageIdentifier'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
        "502":
          description: Bad Gateway
          schema:
            type: string
      summary: submit the message
      tags:
      - RMB
  /twin/{twin_id}/{retqueue}:
    get:
      consumes:
      - application/json
      description: Get the message result
      parameters:
      - description: twin id
        in: path
        name: twin_id
        required: true
        type: integer
      - description: message retqueue
        in: path
        name: retqueue
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/rmbproxy.Message'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
        "502":
          description: Bad Gateway
          schema:
            type: string
      summary: Get the message result
      tags:
      - RMB
...
swagger: "2.0"